Minor correction in output explanation
According to the comment in the source code, the fifth row of the "GENERIC.log" file means the number of tasks executed per layer, instead of the number of fog nodes. Note that the sum of the fifth row is not equal to completed tasks + failed tasks. Further investigation is required.